moved data reading to new file
