+++ main.py
@@ -1,29 +1,9 @@
 import numpy as np
 from graph import *
 from ml import *
+from data import *
 
 print("Tensorflow version "+tf.__version__)
-
-
-def preprocess(func):
-    def wrapper(*args, **kwargs):
-        # load data
-        ret = func(*args, **kwargs)
-        # scale train/test images to 0-1 values
-        ret[0] = ret[0] / 255.0
-        ret[2] = ret[2] / 255.0
-        return tuple(ret)
-    return wrapper
-
-
-@preprocess
-def load_data():
-    fashion_mnist = keras.datasets.fashion_mnist
-    (train_images, train_labels), (test_images, test_labels) = fashion_mnist.load_data()
-    class_names = ['T-shirt/top', 'Trouser', 'Pullover', 'Dress', 'Coat',
-                    'Sandal', 'Shirt', 'Sneaker', 'Bag', 'Ankle boot']
-    return [train_images, train_labels, test_images, test_labels, class_names]
-
 
 if __name__ == "__main__":
